Refactor general analysis code
