+++ soils/analysis/statistics.py
@@ -1,0 +1,18 @@
+import logging
+
+from soils.analysis.compute_general_statistics import statistics
+
+logger = logging.getLogger()
+logger.setLevel(logging.INFO)
+
+
+# def analysis(event, context): #only for lambda
+def analysis(event):
+    logger.info(f'## EVENT\r {event}')
+    # logger.info(f'## CONTEXT\r {context}')
+    request = event  # .get_json()
+
+    dataset = request['dataset']
+    logger.info(f'## DATASET\r {dataset}')
+
+    return statistics(request)
