+++ soils/analysis/utils.py
@@ -1,0 +1,12 @@
+import numpy as np
+from affine import Affine
+
+
+def transform_from_latlon(lat, lon):
+    lat = np.asarray(lat)
+    lon = np.asarray(lon)
+    trans = Affine.translation(lon[0], lat[0])
+    scale = Affine.scale(lon[1] - lon[0], lat[1] - lat[0])
+    return trans * scale
+
+
